+++ CHANGELOG.md
@@ -5,3 +5,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### added
+- Restricts return pickings from pickings of sales orders whose time has expired.